Rename spec and template directories
