+++ cyanobyte-templates/circuitpython.py
@@ -1,5 +1,5 @@
 {% import 'macros.jinja2' as utils %}
-{% import 'python.jinja2' as py %}
+{% import 'base.python.jinja2' as py %}
 {% set template = namespace(sign=false, math=false, struct=false) %}
 {{ utils.pad_string('# ', utils.license(info.copyright.name, info.copyright.date, info.license.name)) -}}
 #
