+++ cyanobyte-templates/cmsis.svd
