+++ cyanobyte-templates/esp32.options.yaml
