+++ cyanobyte-templates/kubos.h
@@ -1,5 +1,5 @@
 {% import 'macros.jinja2' as utils %}
-{% import 'clang.jinja2' as cpp %}
+{% import 'base.c.jinja2' as cpp %}
 {% set template = namespace(math=false) %}
 /*
 {{ utils.pad_string('* ', utils.license(info.copyright.name, info.copyright.date, info.license.name)) -}}
