+++ cyanobyte-templates/mcu.jinja2
