+++ cyanobyte.egg-info/SOURCES.txt
@@ -3,6 +3,7 @@
 .gitmodules
 CONTRIBUTING.md
 LICENSE
+MANIFEST.in
 README.md
 cloudbuild-deploy.yaml
 cloudbuild.yaml
@@ -18,6 +19,32 @@ cyanobyte/__init__.py
 cyanobyte/codegen.py
 cyanobyte/convert_json_to_yaml.py
 cyanobyte/validator.py
+cyanobyte-spec/__init__.py
+cyanobyte-spec/cyanobyte.schema.json
+cyanobyte-templates/__init__.py
+cyanobyte-templates/arduino.cpp
+cyanobyte-templates/arduino.h
+cyanobyte-templates/base.c.jinja2
+cyanobyte-templates/base.javascript.jinja2
+cyanobyte-templates/base.python.jinja2
+cyanobyte-templates/circuitpython.py
+cyanobyte-templates/cmsis.svd
+cyanobyte-templates/datasheet.tex
+cyanobyte-templates/doc.md
+cyanobyte-templates/esp32.options.yaml
+cyanobyte-templates/espruino.js
+cyanobyte-templates/i2c-device.py
+cyanobyte-templates/kubos.c
+cyanobyte-templates/kubos.h
+cyanobyte-templates/macros.jinja2
+cyanobyte-templates/mcu.c
+cyanobyte-templates/mcu.h
+cyanobyte-templates/mcu.jinja2
+cyanobyte-templates/micropython.py
+cyanobyte-templates/raspberrypi-spi-emboss.py
+cyanobyte-templates/raspberrypi-spi-register.py
+cyanobyte-templates/raspberrypi.py
+cyanobyte-templates/webpage.html
 cyanobyte.egg-info/MANIFEST.in
 cyanobyte.egg-info/PKG-INFO
 cyanobyte.egg-info/SOURCES.txt
@@ -77,32 +104,6 @@ peripherals/example.yaml
 peripherals/examplespi-emboss.yaml
 peripherals/examplespi-register.yaml
 peripherals/float.emb
-spec/__init__.py
-spec/cyanobyte.schema.json
-templates/__init__.py
-templates/arduino.cpp
-templates/arduino.h
-templates/circuitpython.py
-templates/clang.jinja2
-templates/cmsis.svd
-templates/datasheet.tex
-templates/doc.md
-templates/esp32.options.yaml
-templates/espruino.js
-templates/generic.c
-templates/generic.h
-templates/generic.jinja2
-templates/i2c-device.py
-templates/javascript.jinja2
-templates/kubos.c
-templates/kubos.h
-templates/macros.jinja2
-templates/micropython.py
-templates/python.jinja2
-templates/raspberrypi-spi-emboss.py
-templates/raspberrypi-spi-register.py
-templates/raspberrypi.py
-templates/webpage.html
 test/__init__.py
 test/pylintrc
 test/test_cmsis_svd.py
