+++ cyanobyte.egg-info/top_level.txt
@@ -1,3 +1,3 @@
 cyanobyte
-spec
-templates
+cyanobyte-spec
+cyanobyte-templates
